requirements.txt
amqp==2.6.0
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==0.24.0
attrs==19.1.0
beautifulsoup4==4.6.0
billiard==3.6.3.0
black==19.3b0
boto3==1.14.29
botocore==1.17.29
cachetools==4.1.0
celery==4.4.6
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
datefinder==0.7.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-braces==1.14.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-graphql-jwt==0.3.1
django-import-export==2.3.0
django-letsencrypt==3.0.1
django-modelcluster==5.0.1
django-nine==0.2.3
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-sslserver==0.22
django-storages==1.9.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
draftjs-exporter==2.1.7
ecdsa==0.15
elasticsearch==6.8.1
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
gspread==3.4.2
hexdump==3.3
html5lib==1.0.1
httplib2==0.17.2
icalendar==4.0.4
idna==2.8
importlib-metadata==1.7.0
intelhex==2.2.1
ipaddress==1.0.23
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
jose==1.0.0
kiwisolver==1.1.0
kombu==4.6.11
l18n==2018.5
lxml==4.5.0
mailchimp3==3.0.12
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.3
multitasking==0.0.9
mysql-connector-python==8.0.19
mysqlclient==1.4.2.post1
numpy==1.18.1
oauth2-provider==0.0
oauth2client==4.1.3
oauthlib==3.0.1
odfpy==1.4.1
openpyxl==3.0.4
pandas==1.0.3
pandas-datareader==0.8.1
phonenumbers==8.11.3
Pillow==6.2.2
promise==2.3
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.6
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.0
python-decouple==3.3
python-jose==2.0.2
python3-openid==3.1.0
pytz==2019.1
pyxero==0.9.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Rx==1.6.1
s3transfer==0.3.3
signrequest-python-client==2.1.0
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.0
stripe==2.42.0
tablib==2.0.0
toml==0.10.0
twilio==6.35.4
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
Willow==1.3
xlrd==1.2.0
xlwt==1.3.0
xmlr==0.3.1
yfinance==0.1.54
zipp==3.1.0
